+++ src/module/data/actor/base-actor.mjs
@@ -3,6 +3,8 @@ import { CombatDialogHelper } from '../../helpers/combat/combat-dialog.mjs';
 import { FoundryAdapter } from '../../helpers/foundry-adapter.mjs';
 import { ChatMessageBuilder } from '../../helpers/ui/chat-message-builder.mjs';
 import { Sanitizer } from '../../helpers/sanitizer.mjs';
+import { SkillLoader } from '../../helpers/character/skill-loader.mjs';
+import { ModifierCollector } from '../../helpers/character/modifier-collector.mjs';
 
 const { fields } = foundry.data;
 
@@ -12,6 +14,60 @@ const { fields } = foundry.data;
  * @extends {DeathwatchDataModel}
  */
 export default class DeathwatchActorBase extends DeathwatchDataModel {
+
+  /**
+   * Prepare base characteristics shared by all actor types.
+   * Handles skill loading, modifier collection/application, and movement calculation.
+   * Returns itemsArray and allModifiers for actor-specific processing.
+   *
+   * @returns {{itemsArray: Item[], allModifiers: Object[]}} Items and collected modifiers
+   * @protected
+   */
+  _prepareBaseCharacteristics() {
+    const actor = this.parent;
+
+    // Load skills from skills registry
+    if (this.skills) {
+      this.skills = SkillLoader.loadSkills(this.skills);
+    }
+
+    // Convert items collection to array
+    const itemsArray = this._getItemsArray();
+
+    // Collect all modifiers from items, talents, traits, etc.
+    const allModifiers = ModifierCollector.collectAllModifiers(actor, itemsArray);
+
+    // Apply characteristic modifiers
+    ModifierCollector.applyCharacteristicModifiers(this.characteristics, allModifiers);
+
+    // Apply skill modifiers
+    if (this.skills) {
+      ModifierCollector.applySkillModifiers(this.skills, allModifiers);
+    }
+
+    // Apply combat-related modifiers
+    this.initiativeBonus = ModifierCollector.applyInitiativeModifiers(allModifiers);
+    ModifierCollector.applyWoundModifiers(this.wounds, allModifiers);
+    ModifierCollector.applyFatigueModifiers(this.fatigue, this.characteristics?.tg?.mod || 0);
+
+    // Apply armor modifiers
+    ModifierCollector.applyArmorModifiers(itemsArray, allModifiers);
+    this.naturalArmorValue = ModifierCollector.calculateNaturalArmor(allModifiers, itemsArray);
+
+    // Apply weapon own modifiers (after characteristics computed)
+    for (const item of itemsArray) {
+      if (item.type === 'weapon') {
+        item.system._applyOwnModifiers();
+      }
+    }
+
+    // Calculate movement from Agility Bonus
+    const agBonus = this.characteristics?.ag?.mod || 0;
+    if (!this.movement) this.movement = {};
+    ModifierCollector.applyMovementModifiers(this.movement, agBonus, allModifiers);
+
+    return { itemsArray, allModifiers };
+  }
 
   /**
    * Convert actor items collection to array.
@@ -27,6 +83,27 @@ export default class DeathwatchActorBase extends DeathwatchDataModel {
       : Array.isArray(actor.items)
         ? actor.items
         : Array.from(actor.items);
+  }
+
+  /**
+   * Apply psyker-specific modifiers (Psy Rating and Force Weapons).
+   * Used by Character and Enemy actors that can be psykers.
+   *
+   * @param {Item[]} itemsArray - Actor's items
+   * @param {Object[]} allModifiers - Collected modifiers
+   * @protected
+   */
+  _applyPsykerModifiers(itemsArray, allModifiers) {
+    if (!this.psyRating) return;
+
+    ModifierCollector.applyPsyRatingModifiers(this.psyRating, allModifiers);
+
+    // Apply force weapon modifiers after psy rating is computed
+    for (const item of itemsArray) {
+      if (item.type === 'weapon') {
+        item.system.applyForceWeaponModifiers();
+      }
+    }
   }
 
   /**
